pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-sync-tool"
version = "1.0.0"
description = "GitHub 仓库同步工具"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "dev@example.com"}
]
keywords = ["git", "github", "sync", "tool"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]

dependencies = [
    "PySide6>=6.6.0",
    "GitPython>=3.1.40",
    "PyGithub>=2.1.1",
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.1",
    "watchdog>=3.0.0",
    "aiosqlite>=0.19.0",
    "PyYAML>=6.0.1",
    "loguru>=0.7.2",
    "python-dateutil>=2.8.2",
    "click>=8.1.7",
    "rich>=13.7.0",
    "Jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-qt>=4.2.0",
]
build = [
    "pyinstaller>=6.3.0",
]

[project.scripts]
git-sync = "src.cli.main:main"

[project.gui-scripts]
git-sync-gui = "src.ui.main_window:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
